requirements.txt
Pillow==7.0.0
requests==2.23.0
